+++ mingling_core/src/program/setup.rs
@@ -3,6 +3,10 @@ use crate::{ProgramCollect, program::Program};
 mod basic;
 pub use basic::*;
 
+#[doc(hidden)]
+pub mod exit_code_control;
+pub use exit_code_control::ExitCodeSetup;
+
 #[cfg(feature = "general_renderer")]
 mod general_renderer;